- Start work on the OO TextFrame. - Will integrate them with the UI later, and do the same thing for shapes.

+++ src/nsp/server/image/Comparer.java
@@ -33,7 +33,7 @@ public class Comparer {
 				}
 			}
 		}
-		return m11 / (m01 + m10 + m11);
+		return (double)m11 / (m01 + m10 + m11);
 	}
 	
 	public boolean isAlike(double score) {
